docs
A test case for a junior python developer. The test assignment required to create an api counter that would use anation to create additional fields and calculations on the database side, so as not to overload the working code. An additional requirement was to cover the project with tests, as well as to implement the ability to filter statistics, as well as sorting.

To run this api you will need to install in a viral environment:

'django',
'djangorestframework',
'django-filte'